Make declarations in header match implemenations
Fixes #2789

+++ sope-xml/SaxObjC/SaxContentHandler.h
@@ -50,8 +50,8 @@
 
 /* CDATA */
 
-- (void)characters:(unichar *)_chars          length:(int)_len;
-- (void)ignorableWhitespace:(unichar *)_chars length:(int)_len;
+- (void)characters:(unichar *)_chars          length:(NSUInteger)_len;
+- (void)ignorableWhitespace:(unichar *)_chars length:(NSUInteger)_len;
 
 /* PIs */
